feat: primitive support for eval
A standard_env has been created that should store inbuilt functions. For
now it only implements add.

An execute function that takes a Vec<Expr> will execute any function
that exists in the standard_env taking the first element in the list as
the function.

Evaluation the parsed ast still has to be implemented.