refactor: add external library for tests, etc.
For writing external tests and accessing internals as a library, this
commit refactors the helper functions into a `lib.rs`.
Instead of running the code inside main, a helper function
`evaluate_expression` takes the code and returns
`Result<Expr, RuntimeError>` for use in tests inside the `tests/`
directory and maybe future external use.