refactor: make Env a new struct, pass env to evaluate_expression
Instead of just a type alias, `Env` is now a new struct. This allows us
to define custom methods on the type. Added methods to it:
- `Env::new()`: initializes env with standard functions and vars.
- `contains_key`: same as HashMap
- `index`: allows indexing (RHS only, not a mutable index)

Made `extract_num` public since it's needed for tests.

Moved lib functions out of `funcs` module (what was that anyway x).

Renamed `read_from_tokens` to `parse` (since it takes a list of tokens
and converts it into an AST).

Renamed `atom` to `parse_int` (because that is what it does).

In tests, removed `extract_number_from_expr` and used `extract_num`
instead. Replaced match with `if let`.