fix: use VecDeque in parser+tokenizer.
VecDeque is a better choice in this case because we do a lot of popping
from the front. VecDeque is faster for that, whereas Vec does it in
O(n).

Using VecDeque also combines removing the first element and checking if
the vec is empty. `pop_front` returns an Option.

Also added some comments explaining parser.