feat: make tokenizer better - better strings, line nos, etc.
The new tokenizer is just a bunch of `match`es and `loop`s. It can
handle:
- Parentheses
- Spaces and newlines
- Strings
- Everything else is an identifier (symbol)

Instead of String, the tokenizer gives a Token which keeps track of line
number, column number (both start and end) and also stores a reference
to the original text (which can be used to print errors, etc.)

As of now, the other fields aren't used (so you will see many warnings).

This fixes the issue where strings could not have spaces in them.

Though there is one issue: the number token does not ensure that there's
only one "." in the string.