refactor: make execute smaller and simpler
Moved lambda execution into `eval/functions.rs` as lambda definition
is also present there. Moves a lot of code from `execute`.

Handle symbol in match instead of separate case for it. Retrieving
symbol from env was already being done when a symbol was found, but
this was only handled for function call.
The same code was repeated in the special case when exprs has only one
symbol. So removed that and added a check in the match body instead -
if there are more exprs, it is considered a func call otherwise it's
just retrieving a symbol.