refactor: add sep. execute_single, add get_symbol in env
Added a new helper in env - `get_symbol` that will try to get the
symbol stored for the given string or an error saying the symbol was
not found.

Added a separate `execute_single` function used to execute a single
`Expr`. This was added because `execute` was used in a lot of places
with a new single-element vec.

`execute` still handles cases where more than one `Expr`s are needed,
but it calls `execute_single` when nothing matches. `execute_single`
has enumerated all cases in match, so it should give a compiler error
when something hasn't been implemented.

Changed `execute` to take a mutable vector since one case can be
a little optimized if it is. When the first arg is a Symbol and that
symbol is a function, we can just replace the first arg with the
symbol and call execute again. Avoids insert at 0.

Oof this commit message is too big.