feat(parser): add TokenTypes and use them for faster parsing
Since we already know the type of token in the tokenizer, we do not have
to duplicate the same string comparisons in the parser.

Made a TokenType enum and stored the appropriate type when returning
token in tokenizer. The parser matches on this enum instead of string
comparisons.

parse_atom is also much simpler as it does not need to re-tokenizer
string.

This improved performance by 50-90% in large test cases.