feat: add program name and struct
Created a new struct `Program` that stores the name and program text.
- also stores text split by lines

Splitting by lines is done in evaluate_expression instead of tokenizer.
- tokenizer now only takes a `Program` as input

`Token` now stores a reference to a `Program` instead of just the vec of
lines.
- because we also need to print the program name when printing an error

REPL now keeps track of the input number.
- Increments it after every input.
- `input[num]` is used as the program name for programs written in the
  REPL.

`display_and_mark` now works on a program-level:
- separate output for each program
- separate output for each line as before